fix some more 64bit warnings/issues
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3799 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/common.c
@@ -1424,7 +1424,7 @@ static int read_rom_data(struct rom_load_data *romdata, const struct RomModule *
 			return 0;
 		numbytes -= bytesleft;
 
-		debugload("  Copying to %08X\n", (int)base);
+		debugload("  Copying to %08X\n", (size_t)base);
 
 		/* unmasked cases */
 		if (datamask == 0xff)
